Clean up opengl sample.
